Update 10_PrepBISF_AV-WinDefend.ps1
Remove erroneous $ sign from Show-BISFProgressBar invocation.

The erroneous $ sign is causing the underlying Show-BISFProgressBar to not be able to check the state of the invoked, current user Windows Defender AV process.
Removing it ensures that when the AV scan invoked under the current user finishes, it will actually move on to the rest of the preparation tasks.